+++ novlog.js
@@ -1,0 +1,9 @@
+function log(req) {
+  console.log(
+    `[${req.ip}] ${req.method} ${req.originalUrl} with body ${JSON.stringify(
+      req.body
+    )}`
+  );
+}
+
+module.exports = { log };